requirements.txt
pandas==2.3.1
SQLAlchemy==2.0.43
psycopg[binary]
pyarrow==21.0.0
apache-airflow==3.0.4
